+++ Application/Global.php
@@ -3,7 +3,7 @@
  * Skytells PHP Framework --------------------------------------------------*
  * @category   Web Development ( Programming )
  * @package    Skytells PHP Framework
- * @version 1.2.2
+ * @version 1.3.2
  * @license Freeware
  * @copyright  2007-2017 Skytells, Inc. All rights reserved.
  * @license    https://www.skytells.net/us/terms  Freeware.
@@ -13,6 +13,7 @@
 
     $startScriptTime=microtime(TRUE);
     require_once("Misc/Settings.php");
+    require_once("Misc/Config/Firewall.php");
     global $Settings;
     ini_set('zlib.output_compression_level', $Settings['GZIP_COMPRESSION_LEVEL']);
     if ($Settings['ENABLE_COMPRESSION'] == true) { if(!ob_start("ob_gzhandler")) ob_start(); }else { ob_start(); }