+++ Application/Library/Classes/Loader.php
@@ -3,13 +3,14 @@
  * Skytells PHP Framework --------------------------------------------------*
  * @category   Web Development ( Programming )
  * @package    Skytells PHP Framework
- * @version 1.2.2
+ * @version 1.3.2
  * @license Freeware
  * @copyright  2007-2017 Skytells, Inc. All rights reserved.
  * @license    https://www.skytells.net/us/terms  Freeware.
  * @author Dr. Hazem Ali ( fb.com/Haz4m )
  * @see The Framework's changelog to be always up to date.
  */
+
    Class Loader
     {
       public function __construct() {
@@ -93,14 +94,15 @@
                            if (!is_array($args) && $args != false && $args != true){ $newName = $args; $args = false; }
 
                             $clName = $this->getClassNameFromFile($path.$File); $OwnerObject = $clName;
-
+                            $namespace = $this->getClassNamespaceFromFile($path.$File);
+                            $realClassName = (class_exists($namespace."\\".$clName)) ? $namespace."\\".$clName : $clName;
                             if (!empty($newName)) { $OwnerObject = $newName;  }
 
                             if ($args != false && is_array($args)){
-                              $refClass = new ReflectionClass($clName);
+                              $refClass = new ReflectionClass($realClassName);
                               $setOwner->$OwnerObject = $refClass->newInstanceArgs($args);
                               }
-                            else { $setOwner->$OwnerObject = new $clName; }
+                            else {  $setOwner->$OwnerObject = new $realClassName(); }
                           }
                           // ----------------------------------------------------------------
 
@@ -139,14 +141,15 @@
                        if (!is_array($args) && $args != false && $args != true){ $newName = $args; $args = false; }
 
                         $clName = $this->getClassNameFromFile($path.$File); $OwnerObject = $clName;
-
+                        $namespace = $this->getClassNamespaceFromFile($path.$File);
+                        $realClassName = (class_exists($namespace."\\".$clName)) ? $namespace."\\".$clName : $clName;
                         if (!empty($newName)) { $OwnerObject = $newName;  }
 
                         if ($args != false && is_array($args)){
-                          $refClass = new ReflectionClass($clName);
+                          $refClass = new ReflectionClass($realClassName);
                           $setOwner->$OwnerObject = $refClass->newInstanceArgs($args);
                           }
-                        else { $setOwner->$OwnerObject = new $clName; }
+                        else {  $setOwner->$OwnerObject = new $realClassName(); }
                       }
                       // ----------------------------------------------------------------
 
@@ -180,14 +183,15 @@
                        if (!is_array($args) && $args != false && $args != true){ $newName = $args; $args = false; }
 
                         $clName = $this->getClassNameFromFile($path.$File); $OwnerObject = $clName;
-
+                        $namespace = $this->getClassNamespaceFromFile($path.$File);
+                        $realClassName = (class_exists($namespace."\\".$clName)) ? $namespace."\\".$clName : $clName;
                         if (!empty($newName)) { $OwnerObject = $newName;  }
 
                         if ($args != false && is_array($args)){
-                          $refClass = new ReflectionClass($clName);
+                          $refClass = new ReflectionClass($realClassName);
                           $setOwner->$OwnerObject = $refClass->newInstanceArgs($args);
                           }
-                        else { $setOwner->$OwnerObject = new $clName; }
+                        else {  $setOwner->$OwnerObject = new $realClassName(); }
                       }
                       // ----------------------------------------------------------------
                       }else {