+++ Application/Library/Classes/Router.php
@@ -3,7 +3,7 @@
  * Skytells PHP Framework --------------------------------------------------*
  * @category   Web Development ( Programming )
  * @package    Skytells PHP Framework
- * @version 1.2.2
+ * @version 1.3.1
  * @license Freeware
  * @copyright  2007-2017 Skytells, Inc. All rights reserved.
  * @license    https://www.skytells.net/us/terms  Freeware.
@@ -123,6 +123,7 @@ Class Router extends Controller {
 
 		return;
 	}
+
 
 	/**
 	 * Reversed routing